Remove `node` exclusion from `webpack` configs
This introduces a change that allows `webpack` to include polyfills for
`node` libraries that otherwise wouldn't run in the browser.

As work progresses we might replace this with updates to `node-simperium`
that don't depend on the `node` libraries. In the meantime this measure
will allow us to proceed in developing changes to the notifications panel
without having to address this deeper issue.